+++ src/plugins/witsec-mailform/functions.php
@@ -3,141 +3,6 @@ use PHPMailer\PHPMailer\PHPMailer;
 use PHPMailer\PHPMailer\Exception;
 use PHPMailer\PHPMailer\SMTP;
 
-// Check if required core files (or lowercase variants) are present
-foreach (["PHPMailer.php", "Exception.php", "SMTP.php"] as $f) {
-	if (file_exists($f))
-		require $f;
-	else {
-		if (file_exists(strtolower($f)))
-			require strtolower($f);
-		else
-			json(false, "E1218", "Required file is missing", $f);
-	}
-}
-
-// All mailform settings
-$to = "{to}";														// To Address
-$toAlt = "{to-alt}";												// Additional Recipients
-$from = "{from}";													// From Address
-$fromName = "{from-name}";											// From Name
-$fromThem = ("{from-them}" == "1" ? true : false);					// Use Sender as From Address
-$fromThemReplyTo = ("{from-them-replyto}" == "1" ? true : false);	// Use Sender as Reply-To
-$fromNameThem = ("{from-name-them}" == "1" ? true : false);			// Use Sender Name as From Name
-$fromNameThemField = "{from-name-them-field}";						// Name of the field(s) that can contain the Sender Name
-$template = "{template}";											// Mail Template
-$autorespondSubjectPrefix = "{autorespond-subjectprefix}";			// Autorespond Form Subject Prefix
-$autorespondSubject = "{autorespond-subject}";						// Autorespond Custom Subject
-$autorespondTemplate = "{autorespond-template}";					// Autorespond Template
-$rcp = ("{recaptcha}" == "1" ? true : false);						// Use reCAPTCHA
-$rcpVersion = "{recaptcha-version}";								// reCAPTCHA Version
-$rcpScore = "{recaptcha-score}";									// reCAPTCHA Score
-$rcpSecret = "{recaptcha-secretkey}";								// reCAPTCHA Secret Key
-$smtp = ("{smtp}" == "1" ? true : false);							// Use SMTP
-$smtpDebug = ("{smtp-debug}" == "1" ? 3 : 0);						// SMTP Debug
-$smtpHost = "{smtp-host}";											// SMTP Host
-$smtpPort = "{smtp-port}";											// SMTP Port
-$smtpSecure = "{smtp-secure}";										// SMTP Use SSL/TLS (empty, ssl or tls)
-$smtpUsername = "{smtp-username}";									// SMTP Username
-$smtpPassword = "{smtp-password}";									// SMTP Password
-$attachmentsMimeTypes = explode(",", "{attachments-mimetypes}");	// Mime Types
-
-
-// Simple status page
-if ($_SERVER["REQUEST_METHOD"] == "GET" && isset($_GET["status"])) {
-	echo "<!doctype html><head><title>Status</title><meta charset='UTF-8'></head><body>";
-
-	// Check PHP version
-	echo (phpversion() >= 7.3 ? "✔️" : "❌") . " PHP Version<br>";
-
-	// Check if OpenSSL is enabled
-	echo (extension_loaded("openssl") ? "✔️" : "❌") . " OpenSSL<br>";
-
-	// Check if function 'file_get_contents' or 'curl_init' exists
-	echo (function_exists("file_get_contents") || function_exists("curl_init") ? "✔️" : "❌") . " Required functions<br>";
-
-	// Check if image functions for CAPTCHA exist
-	echo (function_exists("imagecreatetruecolor") && function_exists("imagecolorallocate") && function_exists("imagejpeg") ? "✔️" : "❌") . " Required functions for CAPTCHA<br>";
-
-	// Check if image functions for CAPTCHA exist
-	echo (function_exists("mime_content_type") ? "✔️" : "❌") . " Required functions for checking file mime types " . (function_exists("mime_content_type") ? "" : "(maybe fileinfo extension is disabled, attachments will still work)") . "<br>";
-
-	// Check if Google can be contacted
-	if (extension_loaded("openssl") && (function_exists("file_get_contents") || function_exists("curl_init"))) {
-		$s = SendReCaptcha("12345", "67890", true);
-		echo ($s ? "✔️" : "❌") . " Connection to Google reCAPTCHA service<br>";
-	} else
-		echo "❌ Connection to Google reCAPTCHA services (missing OpenSSL)<br>";
-
-	// Check connection to SMTP server
-	if ($smtp) {
-		if (function_exists("fsockopen") && extension_loaded("openssl")) {
-			$socket = fsockopen( ($smtpSecure == "ssl" ? "ssl://" : "") . $smtpHost, $smtpPort, $errno, $errstr, 5);
-			echo ($socket ? "✔️" : "❌") . " SMTP Server Connection";
-		} else
-			echo "❌ SMTP Server Connection (missing functions and/or openSSL)<br>";
-	}
-
-	echo "</body></html>";
-	die();
-}
-
-// We only do stuff if there's a POST
-if ($_SERVER["REQUEST_METHOD"] == "POST") {
-	try {
-		// Sanitize user input
-		SanitizeUserInput();
-
-		// reCAPTCHA (only really does anything if reCAPTCHA is enabled)
-		if ($rcp && $rcpVersion != "captcha")
-			CheckReCAPTCHA($rcpScore, $rcpSecret, $rcpVersion);
-
-		// Check captcha
-		if ($rcp && $rcpVersion == "captcha") {
-			session_start();
-			if (isset($_SESSION["captchaCode"]) && isset($_POST["captcha"]) && $_SESSION["captchaCode"] == $_POST["captcha"]) {
-				// Captcha passed!
-			}
-			else
-				json(false, "E1200", "Request did not pass Captcha");
-		}
-
-		// Determine the correct recipient
-		$to = DetermineRecipient($to, $toAlt);
-
-		// Set subject (and take it out of $_POST, so it doesn't end up in the body as well)
-		$subject = $_POST["subject"];
-
-		// Determine if we need to autorespond
-		$autorespond = (isset($_POST["autorespond"]) && $_POST["autorespond"] == "1" ? true : false);
-
-		// Unset some fields that are no longer needed
-		unset($_POST["subject"]);
-		unset($_POST["disableButton"]);
-		unset($_POST["autorespond"]);
-		unset($_POST["redirectURL"]);
-		unset($_POST["captcha"]);
-
-		// Replace any \n breakline with a html breakline
-		$_POST = str_replace("\n", "<br>", $_POST);
-
-		// Send mail (to-address, Sender as From Email or predefined From Email, Sender as From Name or predefined From Name, subject, template)
-		SendPHPMail($to, ($fromThem ? $_POST["email"] : $from), ($fromThemReplyTo ? $_POST["email"] : null), SetFromName(), $subject, $template, false);
-
-		// Send autorespond
-		if ($autorespond) {
-			$arSubject = $autorespondSubject ?: trim($autorespondSubjectPrefix . " " . $subject);
-			SendPHPMail($_POST["email"], $from, null, $fromName, $arSubject, $autorespondTemplate, true);
-		}
-
-		// If we ended up here, it's all good
-		json(true);
-	}
-	catch (Exception $e) {
-		json(false, "E1203", "An unknown error occured");
-	}
-}
-
-
 // Function to return JSON
 function json($success, $errorcode="", $msg="", $args="") {
 	$arr = array("success" => $success);
@@ -235,8 +100,13 @@ function SendReCaptcha($rcpSecret, $rcpResponse, $statusCheck=false) {
 
 // Render Template
 function RenderTemplate($template, $isAutorespond) {
+	global $_SETTINGS;
+
 	// Use a copy of $_POST, so we don't pollute the original
 	$POST = $_POST;
+
+	// Replace BB tags with HTML
+	$template = BBtoHTML($template);
 
 	// Handle "mtxt" and "rtxt"
 	if (!$isAutorespond && isset($POST["mtxt"]))
@@ -271,7 +141,7 @@ function RenderTemplate($template, $isAutorespond) {
 
 	// Add some additional variables to the play
 	$POST["ip"] = GetClientIP(); // Get user's IP address
-	$POST["date"] = ( isset($_POST["date"]) ? $_POST["date"] : date('Y-m-d H:i:s') );	// Only include 'date' if it wasn't used in the form
+	$POST["date"] = ( isset($_POST["date"]) ? $_POST["date"] : date($_SETTINGS["dateFormat"] . " " . $_SETTINGS["timeFormat"]) );	// Only include 'date' if it wasn't used in the form
 
 	// Loop through all variables of the template
 	foreach($matches[1] as $val) {
@@ -287,7 +157,7 @@ function RenderTemplate($template, $isAutorespond) {
 // Use PHPMailer to send the e-mail
 $debug = "";
 function SendPHPMail($to, $from, $replyTo, $fromName, $subject, $template, $isAutorespond) {
-	global $smtp, $smtpDebug, $smtpHost, $smtpPort, $smtpSecure, $smtpUsername, $smtpPassword, $attachmentsMimeTypes, $debug;
+	global $_SETTINGS, $debug;
 
 	try {
 		$mail = new PHPMailer(true);
@@ -308,29 +178,29 @@ function SendPHPMail($to, $from, $replyTo, $fromName, $subject, $template, $isAu
 		}
 
 		// If SMTP
-		if ($smtp) {
+		if ($_SETTINGS["smtp"]) {
 			$mail->isSMTP();
-			$mail->Host = $smtpHost;
-			$mail->Port = $smtpPort;
+			$mail->Host = $_SETTINGS["smtpHost"];
+			$mail->Port = $_SETTINGS["smtpPort"];
 
 			// Debugging
-			if ($smtpDebug) {
-				$mail->SMTPDebug = $smtpDebug;
+			if ($_SETTINGS["smtpDebug"]) {
+				$mail->SMTPDebug = $_SETTINGS["smtpDebug"];
 				$mail->Debugoutput = function($str, $level) {
 					$GLOBALS["debug"] .= "\n" . $str;
 				};
 			}
 
 			// If a username and password are provided
-			if ($smtpUsername && $smtpPassword) {
+			if ($_SETTINGS["smtpUsername"] && $_SETTINGS["smtpPassword"]) {
 				$mail->SMTPAuth = true;
-				$mail->Username = $smtpUsername;
-				$mail->Password = $smtpPassword;
+				$mail->Username = $_SETTINGS["smtpUsername"];
+				$mail->Password = $_SETTINGS["smtpPassword"];
 			}
 
 			// Use SSL/TLS?
-			if ($smtpSecure)
-				$mail->SMTPSecure = $smtpSecure;
+			if ($_SETTINGS["smtpSecure"])
+				$mail->SMTPSecure = $_SETTINGS["smtpSecure"];
 		}
 
 		// Add attachments, if enabled and if there are any
@@ -352,6 +222,9 @@ function SendPHPMail($to, $from, $replyTo, $fromName, $subject, $template, $isAu
 							// Check if the file was uploaded via HTTP POST
 							if (!is_uploaded_file($f["tmp_name"]))
 								json(false, "E1210", "Error uploading file", $f["name"]);
+
+							// Check file extension
+							CheckFileExtension($f);
 
 							// Check mime type
 							CheckMimeType($f);
@@ -382,13 +255,12 @@ function SendPHPMail($to, $from, $replyTo, $fromName, $subject, $template, $isAu
 
 // Set the From Name (this depends on a couple of factors, hence it gets its own function)
 function SetFromName() {
-	global $fromName, $fromNameThem, $fromNameThemField;
-
+	global $_SETTINGS;
 	$fn = "";
 
 	// If we want to use the sender's name as the From Name, parse all name-{fields}
-	if ($fromNameThem) {
-		$fn = $fromNameThemField;
+	if ($_SETTINGS["fromNameThem"]) {
+		$fn = $_SETTINGS["fromNameThemField"];
 
 		// Extract variable names from the name field
 		preg_match_all("/\{([a-zA-Z0-9_-]+)\}/", $fn, $nameMatches);
@@ -402,7 +274,7 @@ function SetFromName() {
 
 	// Double check if fromName isn't empty, otherwise fill it with the predefined name
 	$fn = trim($fn);
-	$fn = ($fn ? $fn : $fromName);
+	$fn = ($fn ? $fn : $_SETTINGS["fromName"]);
 	$fn = preg_replace('/(["“”‘’„”«»]|&quot;)/', "", $fn, -1);
 
 	return $fn;
@@ -411,6 +283,8 @@ function SetFromName() {
 
 // Sanitize user input
 function SanitizeUserInput() {
+	global $_SETTINGS;
+
 	array_walk_recursive($_POST, function(&$input) { $input = strip_tags($input); });
 	array_walk_recursive($_POST, function(&$input) { $input = htmlspecialchars($input); });
 
@@ -426,6 +300,10 @@ function SanitizeUserInput() {
 
 // Determine the correct recipient
 function DetermineRecipient($to, $toAlt) {
+	// If no additional recipients are set, then leave
+	if ($toAlt === "")
+		return $to;
+
 	// Split up additional recipients into an array containing 'identifier' and email address(es)
 	$toAltRec = [];
 	foreach(explode("\n", $toAlt) as $rec) {
@@ -463,9 +341,22 @@ function GetClientIP() {
 }
 
 
+// Check file extension of a file
+function CheckFileExtension($f) {
+	global $_SETTINGS;
+
+	$fext = strtolower(pathinfo($f["name"], PATHINFO_EXTENSION));
+
+	// Let's check if the file extension is part of the allowed extensions array
+	if (!in_array($fext, $_SETTINGS["attachmentsExtensions"])) {
+		json(false, "E1219", "Error uploading file (file extension is not allowed)", $f["name"]);
+	}
+}
+
+
 // Check mime type of a file (if the function is available)
 function CheckMimeType($f) {
-	global $attachmentsMimeTypes;
+	global $_SETTINGS;
 
 	// If function is not available, skip the checks
 	if (!function_exists("mime_content_type"))
@@ -475,9 +366,9 @@ function CheckMimeType($f) {
 
 	// Let's check if the mime type is part of the allowed mime types array
 	$found = false;
-	if (!in_array($fmime, $attachmentsMimeTypes)) {
+	if (!in_array($fmime, $_SETTINGS["attachmentsMimeTypes"])) {
 		// Apparently not, so let's go through all wildcards
-		foreach($attachmentsMimeTypes as $amt) {
+		foreach($_SETTINGS["attachmentsMimeTypes"] as $amt) {
 			if (substr($amt, -1) == "*") {
 				$amt = substr($amt, 0, -1);
 	
@@ -508,5 +399,120 @@ function rearrangeFiles($file) {
 	}
 
 	return $file_ary;
+}
+
+
+// Check if string ends with a specific string (PHP 8 has this built-in, but many still run 7.x)
+function endsWith($haystack, $needle) {
+	$length = strlen($needle);
+	if(!$length)
+		return true;
+	return substr($haystack, -$length) === $needle;
+}
+
+
+// Replace any date/time fields with a formatted date/time
+function FormatDateTime() {
+	global $_SETTINGS;
+
+	foreach ($_POST as $k => $v) {
+		// Check if it's a date/time field
+		$f = "";
+		if (endsWith($k, "-date"))
+			$f = $_SETTINGS["dateFormat"];
+		elseif (endsWith($k, "-time"))
+			$f = $_SETTINGS["timeFormat"];
+		elseif (endsWith($k, "-datetime"))
+			$f = $_SETTINGS["dateFormat"] . " " . $_SETTINGS["timeFormat"];
+
+		// It's a date/time field, so let's format it, replace the original value and rename the key (remove "-date", "-time" and "-datetime")
+		if ($f) {
+			$_POST[$k] = date($f, strtotime($v));
+
+			// Rename current key
+			$newKey = preg_replace("/-(datetime|date|time)$/", "", $k);
+			$_POST = RenameArrayKey($_POST, $k, $newKey);
+		}
+	}
+}
+
+
+// Rename key in associative array
+function RenameArrayKey($arr, $oldkey, $newkey) {
+	if(array_key_exists( $oldkey, $arr)) {
+		$keys = array_keys($arr);
+    	$keys[array_search($oldkey, $keys)] = $newkey;
+	    return array_combine($keys, $arr);	
+	}
+    return $arr;    
+}
+
+
+// Convert allowed BB tags to HTML
+function BBtoHTML($bb) {
+	global $_SETTINGS;
+
+	if ($_SETTINGS["allowedHTML"] !== "") {
+		$tags = explode(",", $_SETTINGS["allowedHTML"]);
+
+		// We need an array
+		$isArr = false;
+		if (!is_array($bb)) {
+			$isArr = true;
+			$bb = (array) $bb;
+		}
+
+		// Replace BB with HTML
+		array_walk_recursive($bb, function(&$input) use ($tags) {
+			foreach ($tags as $tag) {
+				$input = str_replace("[{$tag}]", "<{$tag}>", $input);
+				$input = str_replace("[/{$tag}]", "</{$tag}>", $input);
+			}
+		});
+
+		// Turn back into string
+		if ($isArr)
+			$bb = $bb[0];
+	}
+
+	return $bb;
+}
+
+
+// Format settings
+function FormatSettings() {
+	global $_SETTINGS;
+
+	$_SETTINGS["fromThem"]				= ($_SETTINGS["fromThem"] == "1" ? true : false);
+	$_SETTINGS["fromThemReplyTo"]		= ($_SETTINGS["fromThemReplyTo"] == "1" ? true : false);
+	$_SETTINGS["fromNameThem"]			= ($_SETTINGS["fromNameThem"] == "1" ? true : false);
+	$_SETTINGS["dateFormat"]			= ($_SETTINGS["dateFormat"] ?: "F j, Y");
+	$_SETTINGS["timeFormat"]			= ($_SETTINGS["timeFormat"] ?: "g:i a");
+	$_SETTINGS["rcp"]					= ($_SETTINGS["rcp"] ==  1 ? true : false);
+	$_SETTINGS["smtp"]					= ($_SETTINGS["smtp"] ==  1 ? true : false);
+	$_SETTINGS["smtpDebug"]				= ($_SETTINGS["smtpDebug"] ==  1 ? 3 : 0);
+	$_SETTINGS["attachmentsExtensions"]	= explode(",", $_SETTINGS["attachmentsExtensions"]);
+	$_SETTINGS["attachmentsMimeTypes"]	= explode(",", $_SETTINGS["attachmentsMimeTypes"]);
+}
+
+
+// Replace new lines with <br> (str_replace behaves differently in PHP 8, so can't use str_replace on a multidimensional array anymore)
+function replaceNewlinesWithBr($array) {
+    $newArray = [];
+
+    foreach ($array as $key => $value) {
+        if (is_array($value)) {
+            // Als $value een subarray is, roep de functie opnieuw aan
+            $newArray[$key] = replaceNewlinesWithBr($value);
+        } elseif (is_string($value)) {
+            // Als $value een string is, vervang "\n" door "<br>"
+            $newArray[$key] = str_replace("\n", "<br>", $value);
+        } else {
+            // Als $value geen array of string is, kopieer het gewoon naar de nieuwe array
+            $newArray[$key] = $value;
+        }
+    }
+
+    return $newArray;
 }
 ?>